@@ -1,4 +1,4 @@
-class Agex < ActiveRecord::Base
+class Agex < ApplicationRecord
   belongs_to :user
   belongs_to :vice
 
