@@ -1,5 +1,6 @@
-class Bank < ActiveRecord::Base
+class Bank < ApplicationRecord
   belongs_to :user
+  has_many :accounts
 
   validates :name, presence: true
   validates :access_token, presence: true
