@@ -1,4 +1,4 @@
-class UserVice < ActiveRecord::Base
+class UserVice < ApplicationRecord
   belongs_to :user
   belongs_to :vice
 end
