@@ -1,4 +1,4 @@
-class Vice < ActiveRecord::Base
+class Vice < ApplicationRecord
   has_many :user_vices
   has_many :users, through: :user_vices
 
