@@ -1,8 +1,68 @@
 module UserHelper
+  def set_bank(user, type, access_token)
+    # Find existing Bank or create new one
+    bank = user.banks.find_by(access_token: access_token)
+    unless bank
+      bank = user.banks.new(name: type, access_token: access_token)
+      bank.save!
+    end
+    true
+  end
+
+  def render_mfa_or_populate(user, plaid_user)
+    if plaid_user.mfa?
+      # MFA
+      ret = plaid_user.instance_values.slice 'access_token', 'mfa_type', 'mfa'
+      return render json: ret, status: :ok
+    end
+
+    # No MFA required
+    # Add 'auth' product to get full account numbers
+    begin
+      plaid_user = plaid_user.upgrade(:auth)
+    rescue Plaid::PlaidError => e
+      return handle_plaid_error(e)
+    end
+    ret = populate_user_accounts(user, plaid_user)
+    # 'ret' will either be a successfully saved User model or an ActiveRecord
+    # error hash.
+    unless ret.is_a? User
+      return render json: ret, status: :internal_server_error
+    end
+    render json: ret, status: :ok
+  end
+
+  def handle_plaid_error(e)
+    status = case e
+    when Plaid::BadRequestError
+      :bad_request
+    when Plaid::UnauthorizedError
+      :unauthorized
+    when Plaid::RequestFailedError
+      :payment_required
+    when Plaid::NotFoundError
+      :not_found
+    when Plaid::ServerError
+      :internal_server_error
+    else
+      :internal_server_error
+    end
+    render json: {
+      'code' => e.code,
+      'message' => e.message,
+      'resolve' => e.resolve
+    }, status: status
+  end
+
+  private
+
   def populate_user_accounts(user, plaid_user)
-    # This method returns either a successfully saved User model or an error
-    # hash.
+    bank = user.banks.find_by(access_token: plaid_user.access_token)
+    # This method idempotently populates user's Accounts with accounts given in
+    # plaid_user, then returns either a successfully saved User model or an
+    # error hash.
     plaid_user.accounts.each do |plaid_account|
+      # Get existing Account or create new one
       account = nil
       user.accounts.each do |user_account|
         if plaid_account.id == user_account.plaid_id
@@ -10,27 +70,28 @@ module UserHelper
           break
         end
       end if user.accounts
-      account = user.accounts.new unless account
-      account.plaid_id = plaid_account.id
+      # Create new Account if one wasn't found in loop above
+      unless account
+        account = user.accounts.new
+        account.bank = bank
+        account.plaid_id = plaid_account.id
+        if plaid_account.numbers
+          account.routing_num = plaid_account.numbers[:routing]
+          account.account_num = plaid_account.numbers[:account]
+        else
+          account.account_num = plaid_account.meta['number']
+        end
+      end
+
+      # Populate Account details (or update details if Account already exists)
       account.name = plaid_account.meta['name']
       account.account_type = plaid_account.type.to_s
       account.account_subtype = plaid_account.subtype
       account.institution = plaid_account.institution.to_s
-      if plaid_account.numbers
-        account.routing_num = plaid_account.numbers[:routing]
-        account.account_num = plaid_account.numbers[:account]
-      else
-        account.account_num = plaid_account.meta['number']
-      end
       return account.errors unless account.valid?
       account.save!
     end if plaid_user.accounts
 
     user
-  end
-
-  def set_bank(type, access_token)
-    bank = @authed_user.banks.new(name: type, access_token: access_token)
-    bank.save!
   end
 end
