@@ -1,6 +1,16 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
-require 'rails/all'
+require "rails"
+# Pick the frameworks you want:
+require "active_model/railtie"
+require "active_job/railtie"
+require "active_record/railtie"
+require "action_controller/railtie"
+require "action_mailer/railtie"
+require "action_view/railtie"
+require "action_cable/engine"
+# require "sprockets/railtie"
+require "rails/test_unit/railtie"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -8,28 +18,20 @@ Bundler.require(*Rails.groups)
 
 BCrypt::Engine.cost = 11
 
-module DimentionApp
+module NoncentsApi
   class Application < Rails::Application
-    # Settings in config/environments/* take precedence over those specified
-    # here. Application configuration should go into files in
-    # config/initializers -- all .rb files in that directory are automatically
-    # loaded.
+    # Settings in config/environments/* take precedence over those specified here.
+    # Application configuration should go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded.
+
+    # Only loads a smaller set of middleware suitable for API only apps.
+    # Middleware like session, flash, cookies can be added back manually.
+    # Skip views, helpers and assets when generating a new resource.
+    config.api_only = true
 
     # Set Time.zone default to the specified zone and make Active Record
     # auto-convert to this zone. Run "rake -D time" for a list of tasks for
     # finding time zone names. Default is UTC.
     config.time_zone = 'Pacific Time (US & Canada)'
-
-    # The default locale is :en and all translations from
-    # config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path +=
-    #   Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
-
-    # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
-
-    config.middleware.use Rack::Throttle::Minute, max: 60, code: 420
-    config.middleware.use Rack::Throttle::Hourly, max: 1000, code: 420
   end
 end
