@@ -1,6 +1,10 @@
-# This file should contain all the record creation needed to seed the database
-# with its default values. The data can then be loaded with the rake db:seed
-# (or created alongside the db with db:setup).
+# This file should contain all the record creation needed to seed the database with its default values.
+# The data can then be loaded with the rails db:seed command (or created alongside the database with db:setup).
+#
+# Examples:
+#
+#   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
+#   Character.create(name: 'Luke', movie: movies.first)
 
 vice_names = %w(Nightlife Restaurants Shopping FastFood CoffeeShops Travel
                 Experiences Electronics PersonalCare Movies RideSharing)
@@ -9,7 +13,7 @@ vice_names.each do |name|
 end
 
 user = User.new(fname: 'Cash', lname: 'Money', dob: '1990-01-20',
-                number: '+15555552016', invest_percent: 10,
+                phone: '5555552016', invest_percent: 10,
                 password: 'Ca5hM0n3y', email: 'cashmoney@gmail.com')
 user.create_fund
 user.generate_token
